requirements.txt
django-filter==2.2.0
djangorestframework==3.10.3
django-cors-headers==3.2.0
beautifulsoup4==4.8.1
pytesseract==0.3.0
gunicorn==20.0.3
drf-yasg==1.16.0
requests==2.23.0
psycopg2==2.8.4
cycler==0.10.0
celery==4.3.0
gevent==1.4.0
django==2.2.9
redis==3.3.1
